14. AWS S3 et MiniIO/workshop/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
# lxml==5.1.0
fake-useragent==1.4.0

# Stockage
minio==7.2.3
pymongo==4.6.1

# Utilitaires
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3
tqdm==4.66.1


# Traitement de données
pandas==2.2.0
numpy==1.26.4